backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
flask-PyMongo==2.3.0
werkzeug==3.0.0
